requirements.txt
mysql-connector-python
openpyxl
boto3
pandas